refactor: MoveResult 패키지 변경
